🚚 api: move /assessment-results route to src
