+++ api/lib/application/certification-courses/index.js
@@ -218,30 +218,6 @@ const register = async function (server) {
         tags: ['api'],
       },
     },
-    {
-      method: 'POST',
-      path: '/api/admin/certification-courses/{id}/assessment-results',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.adminMemberHasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: certificationCourseController.updateJuryComments,
-        tags: ['api', 'admin', 'assessment-results', 'certification-courses'],
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs ayant les droits d'accès**\n" +
-            '- Elle permet de mettre à jour les commentaires jury\n' +
-            ' - Elle recrée un assessment result pour mettre à jour les commentaires jury\n',
-        ],
-      },
-    },
   ]);
 };
