+++ api/src/certification/course/application/certification-course-controller.js
@@ -1,5 +1,6 @@
 import { usecases } from '../../shared/domain/usecases/index.js';
 import * as events from '../../../../lib/domain/events/index.js';
+import * as assessmentResultSerializer from '../infrastructure/serializers/jsonapi/assessment-result-serializer.js';
 
 const reject = async function (request, h, dependencies = { events }) {
   const certificationCourseId = request.params.id;
@@ -25,7 +26,21 @@ const unreject = async function (request, h, dependencies = { events }) {
   return h.response().code(200);
 };
 
+const updateJuryComments = async function (request, h, dependencies = { assessmentResultSerializer }) {
+  const certificationCourseId = request.params.id;
+  const deserializedAssessmentResult = await dependencies.assessmentResultSerializer.deserialize(request.payload);
+  const juryId = request.auth.credentials.userId;
+
+  await usecases.updateJuryComments({
+    certificationCourseId,
+    assessmentResult: { ...deserializedAssessmentResult, juryId },
+  });
+
+  return null;
+};
+
 export const certificationCourseController = {
   reject,
   unreject,
+  updateJuryComments,
 };
